ecommerce/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.1.3
django-environ==0.4.5
django-jet==1.0.8
django-simple-history==3.0.0
djangorestframework==3.11.0
filelock==3.4.0
importlib-metadata==4.8.2
pbr==5.8.0
Pillow==8.4.0
platformdirs==2.4.0
psycopg2-binary==2.8.6
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
stevedore==3.5.0
typing-extensions==3.10.0.0
virtualenv==20.10.0
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
zipp==3.6.0
